refactor(locales): move i18n stuff to an exclusive component
